Work with slider

+++ Assets/Scripts/Core/Wheel/ITravelable.cs
@@ -2,6 +2,6 @@ namespace Core.Wheel
 {
     public interface ITravelable
     {
-        public float DistanceTraveled { get; }
+        public int DistanceTraveled { get; }
     }
 }
